Perform CopyAndLinkFiles outside job pool.
Previously the creation of linkfiles and copyfiles would happen in an undefined
order when using -j > 1 (multiprocessing job pool with more than one worker).
This makes the behavior of linkfiles or copyfiles that target a destination
within another project unstable since the other project may not be created
fully when the copy/link is created.

Instead we wait until the job pool completes the initial checkout and then
perform the copy/links synchronously. This should not incur a significant
performance degredation since many jobs would all be I/O system-bound.

Change-Id: I6d7e71a2be4083591e498b56df0b0dd4e7f5cd50