Revert "sync: dropped "NOTICE: --use-superproject is in beta ..." message."
This reverts commit d53cb9549a0c57939b12c5c35a6b581aec2ca36f. As long as repo's reference docs
treat this feature as a WIP and don't cover it well enough to allow all repo maintainers to easily
support it, it is inconsistent to report to users that it is no longer in beta. Thanks for vapier@
for noticing. https://crbug.com/gerrit/15527 tracks the required docs changes before we'd be ready
to roll forward again.

Change-Id: Ic9bd951cfb3c1abf6e1bfa30dfe4afa1c9b7bec6