init: add an option --enable-git-lfs-filter
It was reported that git-lfs did not work with git-repo.
Specifically, `git read-tree -u` run by `repo sync` would
fail git-lfs's smudge filter.
See https://github.com/github/git-lfs/issues/1422.

In fact, by the time `git read-tree -u` is run, the repository
is not bare. It is just that, the working directory is not the
same as the .git directory. git-lfs's filter should work. No one
seems to have delved into that issue.

Today, with newer versions of git-repo and git-lfs, that issue
will not reproduce. Tested with
- git 2.33, git-lfs 2.13 on macOS
- git 2.17, git-lfs 2.3 on ubuntu

So, it seems fine to add an option --enable-git-lfs-filter,
default to false, and stat that it may not work with older
versions of git and git-lfs in the help doc.

Change-Id: I8d21854eeeea541e072f63d6b10ad1253b1a9826
Signed-off-by: XD Trol <milestonejxd@gmail.com>