forall: make windows can use environment variable
When running the following command in windows:
repo forall -c 'echo %REPO_PATH%'

The parameter of 'shell' for subprocess will be True.
The command which send to subprocess will be:
['echo %REPO_PATH%', 'echo %REPO_PATH%']

Which is not working on windows.
This change makes it become: 'echo $REPO_PATH' when shell is True.

But block it when user passing additional arguments.
Since windows doesn't support running command_string & passing
arguments to shell like in Linux:
Ex:
$ /bin/sh -c 'printf "%s\n" $@'  cmd_name   1st_arg 2end_arg
1st_arg
2end_arg

The following command won't work in windows:
cmd.exe /c '"printf \"%s\n\" %*" arg1 arg2'

And so far can't find an equivalent way for the same purpose.

Change-Id: Ia74e3f671f0c3ddebf48d438f7ab0714fe6394c3