command: Set default chunksize to 1
A high chunksize makes sense for quick operations to save on IPC
overhead, but having it set for all operations may have negative effect.

For example, let's assume subcmd sync has chunksize of 32, number of
jobs is set to 64 and total number of projects to sync less than 2048.
This means that all jobs will be prescheduled. It's very much possible
that one process gets many slow operations that will be executed
sequentially while other processes finished they work and are idling.

chunksize of 32 was originally chosen for branches subcmd and it was
only used there [0]. With a refactoring that happend afterwards, that
chunksize was moved to command.py [1].

This change sets default chunksize to 1, and restores 32 for branch operation.

[0] https://gerrit-review.googlesource.com/c/git-repo/+/291342
[1] https://gerrit-review.googlesource.com/c/git-repo/+/297242

Change-Id: I9acfbf580542dfe85a03ac0ffad5c8a95f5e2be2