Windows: Use MORE.COM as the default pager
More is the default pager that ships with Windows. It is usually
available on the path.

We will use it if it is available, otherwise we will revert to using
`less`, which users will have to install separately.

The pager will be opened with a utf-8 encoded pipe to make sure the
proper types are passed on.

Now `repo help` will no longer print `fatal: cannot start pager "less"`
if less is not installed.

Change-Id: I3cfe46bb20a7fdccb7db597d0259179da05cfb52