Update errors to extend BaseRepoError
In order to better analyze and track repo errors, repo command failures
need to be tied to specific errors in repo source code.

The first step in this process is a general error refactoring.

This change is the result of a request to break up an existing change into smaller pieces:
- https://gerrit-review.git.corp.google.com/c/git-repo/+/379614

Bug: b/293344017
Change-Id: I46944b1825ce892757c8dd3f7e2fab7e460760c0