manifest_xml: use a set in projectsDiff to speed things up
On a manifests containing many projects, using a set instead
of a sorted list makes the following operations more efficient:
 - membership operation ("in")
 - removal

Also, the sort can be performed on the the remaining elements at the
end (which will usually involve a much smaller number of elements).

(The performance gain is invisible in comparison to the time being
spent performing git operations).

Cosmetic changes:
 - the definition of 'fromProj' is moved to be used in more places
 - the values in diff["added"] are added with a single call to extend

Change-Id: I5ed22ba73b50650ca2d3a49a1ae81f02be3b3055