Use set for faster operations XmlManifest::projectsDiff
On a manifests containing many projects, using a set instead
of a sorted list makes the following operations faster:
 - membership operation ("in")
 - removal

The sort can be performed on the the remaining elements at the
end (which will usually involve a much smaller number of elements).

I took this chance to move the definition of 'fromProj'.

Change-Id: I5ed22ba73b50650ca2d3a49a1ae81f02be3b3055