Allow absolute URLs for remotes
Currently, all fetch URLs for remotes have the path of the manifest
repo pre-pended to them.  This may be desirable in some cases (eg: the
default remote) but this makes it impossible to have remotes that
refer to different hosts.

This change introduces an extra field to the `remote` tag which will
allow us to determine whether to treat the fetch URL as absolute or
not with the default behaviour being the current behaviour of relative
fetch URLs.

Change-Id: I7b79fc4ed276630fdbeb235b94e327b172f0879b