launcher: Use Python 2-compatible string prefix
When commit `b861511` ran Black on the `repo` launcher, it changed the
string prefix in one place from `br` to `rb`. These are synonyms in
Python 3.3 and newer, but the latter form is invalid in older Python
versions[0,1], and in particular not compatible with Python 2.7, making
it break in the bytecode compilation step and not allowing the launcher
to re-launch under Python 3.

`docs/python-support.md` still says that the launcher script should be
compatible with both Python 2 and 3. In addition, since the script
shebang is `#!/usr/bin/env python`, e.g. Debian-based systems will by
default use Python 2 if it is installed (or in modern systems likely
just lack a target altogether, but that is a separate issue).

Change the prefix back to `br`, add a comment on why and a directive for
Black to skip formatting this line in the future.

[0]: https://docs.python.org/3/reference/lexical_analysis.html#string-and-bytes-literals
[1]: https://github.com/psf/black/pull/2297

Change-Id: Iccfc339d2fe8d0c41ea8c649be7e0ad1865d483b