manifest: add optional base check on remove and extend
This adds an optional, built-in checker in git-repo
for making sure patches are not hanging on wrong
base revisions, which is useful if a lower layer of
the manifest changes after a patch was done.

When adding a patch with a new revision using
extend-project or remove-project/project:

          C---D---E patches in project bla
         /
    A---B project bla in manifest state 1

<extend-project name="bla" revision="E" base="B">

If project bla gets updated, in a new snap ID
or by a supplier or similar, to a new state:

          C---D---E patches in project bla
         /
    A---B---F---G project bla in manifest state 2

Parsing will fail because revision of bla
is now G, giving you the choice to create a new
patch branch from G and updating base, or keeping
your previous branch for some reason and just
updating base.

Makes most sense if you use a layered manifest
with includes, and with hashed revisions, but
strings for branches or tags also work fine
when comparing.

Change-Id: Ic6211550a7d3cc9656057f6a2087c505b40cad2b