sync: fix connection error on macos
With large number of sync workers, the sync may fail on macos with
connection errors. The root cause is that multiple workers may connect
to the multiprocessing manager server when handling the first job at
the same time. Then the connection may fail if too many connections
are pending and execeeded the socket listening backlog.

Bug: 377538810
Change-Id: I1924d318d076ca3be61d75daa37bfa8d7dc23ed7