Add Perfetto tracing to repo - DO NOT SUBMIT
This CL adds Perfetto tracing to repo. The tracing is enabled by default

The Perfetto traces can be viewed using the Perfetto UI.

The Perfetto traces can be used to identify performance bottlenecks in repo. For example, the traces can be used to identify the slowest projects to sync, or the slowest commands to run.

Change-Id: Id0e93b229c2a10ea864944d32087713c46900a91