Fix fail in test_repo_trace if trace is disabled
If env has REPO_TRACE 0, then this test produces:
FileNotFoundError: No such file or directory: TRACE_FILE_from_test'
because it assumes trace is always enabled.

Improve by enabling trace for this test if it is disabled.

Modify the repo_trace.SetTrace API to take an optional
setting to be able to also disable trace afterwards
to not impact anything else, still defaults to True.

Change-Id: I2bb4632eb45ee529fc1785833e0bafd9c36f0875