Update internal filesystem layout for submodules
Change the bare checkout directory for submodules from 'subprojects'
to 'modules'. Git expects bare submodule checkouts to be in the
'modules' directory. For consistency, rename 'subproject-objects' to
'module-objects'. If old subproject directories are found, they will
be migrated to the new module directories. This change is the first
step in ensuring Git can understand repo's submodules to some extent.

Change-Id: I385029f1bb55d040616d970d6ffb4bb856692520